# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go modular monolith backend for emergency alert intake, ambulance dispatch, user management, and fleet readiness.",
        "title": "Dispatch Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/dispatch/assignments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated dispatch assignments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "List dispatch assignments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "incident_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ambulance ID",
                        "name": "ambulance_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Assignment status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "status",
                            "assigned_at"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a manual, assisted, or automatic dispatch assignment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Create dispatch assignment",
                "parameters": [
                    {
                        "description": "Dispatch assignment payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_dispatch_application_dto.CreateDispatchAssignmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/assignments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a dispatch assignment by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Get dispatch assignment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dispatch assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/assignments/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates dispatch assignment lifecycle status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Update dispatch assignment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dispatch assignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dispatch status payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_dispatch_application_dto.UpdateDispatchStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/evaluate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uses triage responses to determine whether automatic dispatch should be triggered",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Evaluate automatic dispatch",
                "parameters": [
                    {
                        "description": "Triage dispatch evaluation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_dispatch_application_dto.EvaluateDispatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/recommendations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns generated dispatch recommendations for an incident",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "List dispatch recommendations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Incident ID",
                        "name": "incident_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "generated_at",
                            "score"
                        ],
                        "type": "string",
                        "description": "Sort field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/recommendations/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates scored dispatch recommendations for an incident",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Generate dispatch recommendations",
                "parameters": [
                    {
                        "description": "Recommendation payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_dispatch_application_dto.GenerateRecommendationsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/dispatch/triage": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Saves incident triage responses and derives dispatch priority",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dispatch"
                ],
                "summary": "Persist incident triage",
                "parameters": [
                    {
                        "description": "Persist triage payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_dispatch_application_dto.PersistTriageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dispatch_internal_modules_dispatch_application_dto.CreateDispatchAssignmentRequest": {
            "type": "object",
            "required": [
                "incident_id"
            ],
            "properties": {
                "ambulance_id": {
                    "type": "string"
                },
                "assigned_by_user_id": {
                    "type": "string"
                },
                "assignment_mode": {
                    "type": "string"
                },
                "driver_user_id": {
                    "type": "string"
                },
                "eta_minutes": {
                    "type": "integer"
                },
                "incident_id": {
                    "type": "string"
                },
                "lead_medic_user_id": {
                    "type": "string"
                },
                "ranking_score": {
                    "type": "number"
                }
            }
        },
        "dispatch_internal_modules_dispatch_application_dto.EvaluateDispatchRequest": {
            "type": "object",
            "required": [
                "incident_id",
                "responses"
            ],
            "properties": {
                "incident_id": {
                    "type": "string"
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "question_code",
                            "response_value"
                        ],
                        "properties": {
                            "question_code": {
                                "type": "string"
                            },
                            "response_value": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "dispatch_internal_modules_dispatch_application_dto.GenerateRecommendationsRequest": {
            "type": "object",
            "required": [
                "incident_id"
            ],
            "properties": {
                "auto": {
                    "type": "boolean"
                },
                "incident_id": {
                    "type": "string"
                }
            }
        },
        "dispatch_internal_modules_dispatch_application_dto.PersistTriageRequest": {
            "type": "object",
            "required": [
                "incident_id",
                "questionnaire_code",
                "responses"
            ],
            "properties": {
                "incident_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "questionnaire_code": {
                    "type": "string"
                },
                "responses": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "required": [
                            "question_code",
                            "response_value"
                        ],
                        "properties": {
                            "question_code": {
                                "type": "string"
                            },
                            "response_value": {
                                "type": "string"
                            }
                        }
                    }
                },
                "triage_mode": {
                    "type": "string"
                }
            }
        },
        "dispatch_internal_modules_dispatch_application_dto.UpdateDispatchStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "cancellation_reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}